Increasing modularity and pluggability - `Application` module to multiple modules - step: Refactoring within the module.
`Application` module will have interfaces only.
The extensions to the interfaces will be provided in separate modules: `application-native`, `application-arrow`,....your custom module ;)
