+++ application/src/main/kotlin/com/fraktalio/fmodel/application/MaterializedView.kt
@@ -37,72 +37,16 @@ import kotlinx.coroutines.flow.map
  */
 interface MaterializedView<S, E> : IView<S, E>, ViewStateRepository<E, S> {
     /**
-     * Handles the event of type [E]
-     *
-     * @param event Event of type [E] to be handled
-     * @return [Either] [Error] or State of type [S]
-     */
-    suspend fun handleEither(event: E): Either<Error, S> =
-        // Arrow provides a Monad instance for Either. Except for the types signatures, our program remains unchanged when we compute over Either. All values on the left side assume to be Right biased and, whenever a Left value is found, the computation short-circuits, producing a result that is compatible with the function type signature.
-        either {
-            (event.eitherFetchStateOrFail().bind() ?: initialState)
-                .eitherComputeNewStateOrFail(event).bind()
-                .eitherSaveOrFail().bind()
-        }
-
-    /**
-     * Handles the flow of events of type [E]
-     *
-     * @param events Flow of Events of type [E] to be handled
-     * @return [Flow] of [Either] [Error] or State of type [S]
-     */
-    fun handleEither(events: Flow<E>): Flow<Either<Error, S>> =
-        events
-            .map { handleEither(it) }
-            .catch { emit(Either.Left(Error.EventPublishingFailed(it))) }
-
-    /**
-     * Handles the event of type [E]
-     *
-     * @param event Event of type [E] to be handled
-     * @return State of type [S]
-     */
-    suspend fun handle(event: E): S =
-        (event.fetchState() ?: initialState)
-            .computeNewState(event)
-            .save()
-
-    /**
-     * Handles the flow of events of type [E]
-     *
-     * @param events Flow of Events of type [E] to be handled
-     * @return [Flow] of State of type [S]
-     */
-    fun handle(events: Flow<E>): Flow<S> =
-        events.map { handle(it) }
-
-    /**
      * Computes new State based on the Event.
      *
      * @param event of type [E]
      * @return The newly computed state of type [S]
      */
     fun S.computeNewState(event: E): S = evolve(this, event)
-
-    /**
-     * Computes new State based on the Event or fails.
-     *
-     * @param event of type [E]
-     * @return The newly computed state of type [S] or [Error]
-     */
-    fun S.eitherComputeNewStateOrFail(event: E): Either<Error, S> =
-        Either.catch {
-            computeNewState(event)
-        }.mapLeft { throwable -> Error.CalculatingNewViewStateFailed(this, event, throwable) }
 }
 
 /**
- * Extension function - Materialized View factory function.
+ * Materialized View factory function.
  *
  * The Delegation pattern has proven to be a good alternative to implementation inheritance, and Kotlin supports it natively requiring zero boilerplate code.
  *
@@ -119,6 +63,35 @@ fun <S, E> materializedView(
     viewStateRepository: ViewStateRepository<E, S>,
 ): MaterializedView<S, E> =
     object : MaterializedView<S, E>, ViewStateRepository<E, S> by viewStateRepository, IView<S, E> by view {}
+
+
+// #################################################
+// ##########  Native Kotlin Extensions   ##########
+// #################################################
+
+/**
+ * Extension function - Handles the event of type [E]
+ *
+ * @param event Event of type [E] to be handled
+ * @return State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+suspend fun <S, E> MaterializedView<S, E>.handle(event: E): S =
+    (event.fetchState() ?: initialState)
+        .computeNewState(event)
+        .save()
+
+/**
+ * Extension function - Handles the flow of events of type [E]
+ *
+ * @param events Flow of Events of type [E] to be handled
+ * @return [Flow] of State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <S, E> MaterializedView<S, E>.handle(events: Flow<E>): Flow<S> =
+    events.map { handle(it) }
 
 /**
  * Extension function - Publishes the event of type [E] to the materialized view of type  [MaterializedView]<[S], [E]>
@@ -139,6 +112,75 @@ suspend fun <S, E> E.publishTo(materializedView: MaterializedView<S, E>): S = ma
  * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <S, E> Flow<E>.publishTo(materializedView: MaterializedView<S, E>): Flow<S> = materializedView.handle(this)
+
+
+// #################################################
+// ##########   Arrow Kotlin Extensions   ##########
+// #################################################
+
+/**
+ * Extension function - Handles the event of type [E]
+ *
+ * @param event Event of type [E] to be handled
+ * @return [Either] [Error] or State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+suspend fun <S, E> MaterializedView<S, E>.handleEither(event: E): Either<Error, S> {
+    /**
+     * Inner function - Computes new State based on the Event or fails.
+     *
+     * @param event of type [E]
+     * @return The newly computed state of type [S] or [Error]
+     */
+    fun S.eitherComputeNewStateOrFail(event: E): Either<Error, S> =
+        Either.catch {
+            computeNewState(event)
+        }.mapLeft { throwable -> Error.CalculatingNewViewStateFailed(this, event, throwable) }
+
+    /**
+     * Inner function - Fetch state - either version
+     *
+     * @receiver Event of type [E]
+     * @return [Either] [Error] or the State of type [S]?
+     */
+    suspend fun E.eitherFetchStateOrFail(): Either<Error.FetchingViewStateFailed<E>, S?> =
+        Either.catch {
+            fetchState()
+        }.mapLeft { throwable -> Error.FetchingViewStateFailed(this, throwable) }
+
+    /**
+     * Inner function - Save state - either version
+     *
+     * @receiver State of type [S]
+     * @return [Either] [Error] or the newly saved State of type [S]
+     */
+    suspend fun S.eitherSaveOrFail(): Either<Error.StoringStateFailed<S>, S> =
+        Either.catch {
+            this.save()
+        }.mapLeft { throwable -> Error.StoringStateFailed(this, throwable) }
+
+    // Arrow provides a Monad instance for Either. Except for the types signatures, our program remains unchanged when we compute over Either. All values on the left side assume to be Right biased and, whenever a Left value is found, the computation short-circuits, producing a result that is compatible with the function type signature.
+    return either {
+        (event.eitherFetchStateOrFail().bind() ?: initialState)
+            .eitherComputeNewStateOrFail(event).bind()
+            .eitherSaveOrFail().bind()
+    }
+}
+
+/**
+ * Extension function - Handles the flow of events of type [E]
+ *
+ * @param events Flow of Events of type [E] to be handled
+ * @return [Flow] of [Either] [Error] or State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <S, E> MaterializedView<S, E>.handleEither(events: Flow<E>): Flow<Either<Error, S>> =
+    events
+        .map { handleEither(it) }
+        .catch { emit(Either.Left(Error.EventPublishingFailed(it))) }
+
 
 /**
  * Extension function - Publishes the event of type [E] to the materialized view of type  [MaterializedView]<[S], [E]>
