+++ application/src/main/kotlin/com/fraktalio/fmodel/application/SagaManager.kt
@@ -33,62 +33,18 @@ import kotlinx.coroutines.flow.map
  * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 interface SagaManager<AR, A> : ISaga<AR, A>, ActionPublisher<A> {
-    /**
-     * Handles the action result of type [AR].
-     *
-     * @param actionResult Action Result represent the outcome of some action you want to handle in some way
-     * @return [Flow] of [Either] [Error] or Actions of type [A]
-     */
-    fun handleEither(actionResult: AR): Flow<Either<Error, A>> =
-        actionResult
-            .computeNewActions()
-            .publish()
-            .map { Either.Right(it) }
-            .catch<Either<Error, A>> { emit(Either.Left(Error.ActionResultHandlingFailed(actionResult))) }
-
-    /**
-     * Handles the action result of type [AR].
-     *
-     * @param actionResult Action Result represent the outcome of some action you want to handle in some way
-     * @return [Flow] of Actions of type [A]
-     */
-    fun handle(actionResult: AR): Flow<A> =
-        actionResult
-            .computeNewActions()
-            .publish()
-
-    /**
-     * Handles the the [Flow] of action results of type [AR].
-     *
-     * @param actionResults Action Results represent the outcome of some action you want to handle in some way
-     * @return [Flow] of [Either] [Error] or Actions of type [A]
-     */
-    fun handleEither(actionResults: Flow<AR>): Flow<Either<Error, A>> =
-        actionResults
-            .flatMapConcat { handleEither(it) }
-            .catch { emit(Either.Left(Error.ActionResultPublishingFailed(it))) }
-
-    /**
-     * Handles the the [Flow] of action results of type [AR].
-     *
-     * @param actionResults Action Results represent the outcome of some action you want to handle in some way
-     * @return [Flow] of Actions of type [A]
-     */
-    fun handle(actionResults: Flow<AR>): Flow<A> =
-        actionResults.flatMapConcat { handle(it) }
-
 
     /**
      * Computes new Actions based on the Action Results.
      *
      * @return The newly computed [Flow] of Actions/[A]
      */
-    private fun AR.computeNewActions(): Flow<A> = react(this)
+    fun AR.computeNewActions(): Flow<A> = react(this)
 
 }
 
 /**
- * Extension function - Saga Manager factory function.
+ * Saga Manager factory function.
  *
  * The Delegation pattern has proven to be a good alternative to implementation inheritance, and Kotlin supports it natively requiring zero boilerplate code.
  *
@@ -106,11 +62,41 @@ fun <AR, A> sagaManager(
 ): SagaManager<AR, A> =
     object : SagaManager<AR, A>, ActionPublisher<A> by actionPublisher, ISaga<AR, A> by saga {}
 
+// #################################################
+// ##########  Native Kotlin Extensions   ##########
+// #################################################
+
+/**
+ * Extension function - Handles the action result of type [AR].
+ *
+ * @param actionResult Action Result represent the outcome of some action you want to handle in some way
+ * @return [Flow] of Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <AR, A> SagaManager<AR, A>.handle(actionResult: AR): Flow<A> =
+    actionResult
+        .computeNewActions()
+        .publish()
+
+/**
+ * Extension function - Handles the the [Flow] of action results of type [AR].
+ *
+ * @param actionResults Action Results represent the outcome of some action you want to handle in some way
+ * @return [Flow] of Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <AR, A> SagaManager<AR, A>.handle(actionResults: Flow<AR>): Flow<A> =
+    actionResults.flatMapConcat { handle(it) }
+
 /**
  * Extension function - Publishes the action result of type [AR] to the saga manager of type  [SagaManager]<[AR], [A]>
  * @receiver action result of type [AR]
  * @param sagaManager of type [SagaManager]<[AR], [A]>
  * @return the [Flow] of published Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <AR, A> AR.publishTo(sagaManager: SagaManager<AR, A>): Flow<A> = sagaManager.handle(this)
 
@@ -119,14 +105,52 @@ fun <AR, A> AR.publishTo(sagaManager: SagaManager<AR, A>): Flow<A> = sagaManager
  * @receiver [Flow] of action results of type [AR]
  * @param sagaManager of type [SagaManager]<[AR], [A]>
  * @return the [Flow] of published Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <AR, A> Flow<AR>.publishTo(sagaManager: SagaManager<AR, A>): Flow<A> = sagaManager.handle(this)
+
+
+// #################################################
+// ##########  Arrow Kotlin Extensions   ##########
+// #################################################
+
+/**
+ * Extension function - Handles the action result of type [AR].
+ *
+ * @param actionResult Action Result represent the outcome of some action you want to handle in some way
+ * @return [Flow] of [Either] [Error] or Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <AR, A> SagaManager<AR, A>.handleEither(actionResult: AR): Flow<Either<Error, A>> =
+    actionResult
+        .computeNewActions()
+        .publish()
+        .map { Either.Right(it) }
+        .catch<Either<Error, A>> { emit(Either.Left(Error.ActionResultHandlingFailed(actionResult))) }
+
+/**
+ * Extension function - Handles the the [Flow] of action results of type [AR].
+ *
+ * @param actionResults Action Results represent the outcome of some action you want to handle in some way
+ * @return [Flow] of [Either] [Error] or Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <AR, A> SagaManager<AR, A>.handleEither(actionResults: Flow<AR>): Flow<Either<Error, A>> =
+    actionResults
+        .flatMapConcat { handleEither(it) }
+        .catch { emit(Either.Left(Error.ActionResultPublishingFailed(it))) }
+
 
 /**
  * Extension function - Publishes the action result of type [AR] to the saga manager of type  [SagaManager]<[AR], [A]>
  * @receiver action result of type [AR]
  * @param sagaManager of type [SagaManager]<[AR], [A]>
  * @return the [Flow] of [Either] [Error] or successfully published Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <AR, A> AR.publishEitherTo(sagaManager: SagaManager<AR, A>): Flow<Either<Error, A>> = sagaManager.handleEither(this)
 
@@ -135,6 +159,8 @@ fun <AR, A> AR.publishEitherTo(sagaManager: SagaManager<AR, A>): Flow<Either<Err
  * @receiver [Flow] of action results of type [AR]
  * @param sagaManager of type [SagaManager]<[AR], [A]>
  * @return the [Flow] of [Either] [Error] or successfully published Actions of type [A]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <AR, A> Flow<AR>.publishEitherTo(sagaManager: SagaManager<AR, A>): Flow<Either<Error, A>> =
     sagaManager.handleEither(this)
