+++ application/src/main/kotlin/com/fraktalio/fmodel/application/StateStoredAggregate.kt
@@ -51,64 +51,6 @@ interface StateStoredAggregate<C, S, E> : IDecider<C, S, E>, StateRepository<C, 
         val events = decide(command, this)
         return events.fold(this) { s, e -> evolve(s, e) }
     }
-
-    /**
-     * Computes new State based on the previous State and the [command] or fails.
-     *
-     * @param command of type [C]
-     * @return [Either] the newly computed state of type [S] or [Error]
-     */
-    suspend fun S.eitherComputeNewStateOrFail(command: C): Either<Error, S> =
-        Either.catch {
-            computeNewState(command)
-        }.mapLeft { throwable ->
-            Error.CalculatingNewStateFailed(this, command, throwable)
-        }
-
-    /**
-     * Handles the command message of type [C]
-     *
-     * @param command Command message of type [C]
-     * @return State of type [S]
-     */
-    suspend fun handle(command: C): S =
-        (command.fetchState() ?: initialState)
-            .computeNewState(command)
-            .save()
-
-    /**
-     * Handles the [Flow] of command messages of type [C]
-     *
-     * @param commands [Flow] of Command messages of type [C]
-     * @return [Flow] of State of type [S]
-     */
-    fun handle(commands: Flow<C>): Flow<S> =
-        commands.map { handle(it) }
-
-    /**
-     * Handles the command message of type [C]
-     *
-     * @param command Command message of type [C]
-     * @return Either [Error] or State of type [S]
-     */
-    suspend fun handleEither(command: C): Either<Error, S> =
-        // Arrow provides a Monad instance for Either. Except for the types signatures, our program remains unchanged when we compute over Either. All values on the left side assume to be Right biased and, whenever a Left value is found, the computation short-circuits, producing a result that is compatible with the function type signature.
-        either {
-            (command.eitherFetchStateOrFail().bind() ?: initialState)
-                .eitherComputeNewStateOrFail(command).bind()
-                .eitherSaveOrFail().bind()
-        }
-
-    /**
-     * Handles the [Flow] of command messages of type [C]
-     *
-     * @param commands [Flow] of Command messages of type [C]
-     * @return [Flow] of [Either] [Error] or State of type [S]
-     */
-    fun handleEither(commands: Flow<C>): Flow<Either<Error, S>> =
-        commands
-            .map { handleEither(it) }
-            .catch { emit(Either.Left(Error.CommandPublishingFailed(it))) }
 }
 
 /**
@@ -146,6 +88,29 @@ interface StateStoredOrchestratingAggregate<C, S, E> : ISaga<E, C>, StateStoredA
 }
 
 /**
+ * State stored aggregate factory function.
+ *
+ * The Delegation pattern has proven to be a good alternative to implementation inheritance, and Kotlin supports it natively requiring zero boilerplate code.
+ *
+ * @param C Commands of type [C] that this aggregate can handle
+ * @param S Aggregate state of type [S]
+ * @param E Events of type [E] that are used internally to build/fold new state
+ * @param decider A decider component of type [IDecider]<[C], [S], [E]>
+ * @param stateRepository An aggregate state repository of type [StateRepository]<[C], [S]>
+ * @return An object/instance of type [StateStoredAggregate]<[C], [S], [E]>
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <C, S, E> stateStoredAggregate(
+    decider: IDecider<C, S, E>,
+    stateRepository: StateRepository<C, S>
+): StateStoredAggregate<C, S, E> =
+    object :
+        StateStoredAggregate<C, S, E>,
+        StateRepository<C, S> by stateRepository,
+        IDecider<C, S, E> by decider {}
+
+/**
  * State stored orchestrating aggregate factory function.
  *
  * The Delegation pattern has proven to be a good alternative to implementation inheritance, and Kotlin supports it natively requiring zero boilerplate code.
@@ -170,29 +135,34 @@ fun <C, S, E> stateStoredOrchestratingAggregate(
         IDecider<C, S, E> by decider,
         ISaga<E, C> by saga {}
 
+
+// #################################################
+// ##########  Native Kotlin Extensions   ##########
+// #################################################
+
 /**
- * Extension function - State stored aggregate factory function.
+ * Extension function - Handles the command message of type [C]
  *
- * The Delegation pattern has proven to be a good alternative to implementation inheritance, and Kotlin supports it natively requiring zero boilerplate code.
- *
- * @param C Commands of type [C] that this aggregate can handle
- * @param S Aggregate state of type [S]
- * @param E Events of type [E] that are used internally to build/fold new state
- * @param decider A decider component of type [IDecider]<[C], [S], [E]>
- * @param stateRepository An aggregate state repository of type [StateRepository]<[C], [S]>
- * @return An object/instance of type [StateStoredAggregate]<[C], [S], [E]>
+ * @param command Command message of type [C]
+ * @return State of type [S]
  *
  * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
-fun <C, S, E> stateStoredAggregate(
-    decider: IDecider<C, S, E>,
-    stateRepository: StateRepository<C, S>
-): StateStoredAggregate<C, S, E> =
-    object :
-        StateStoredAggregate<C, S, E>,
-        StateRepository<C, S> by stateRepository,
-        IDecider<C, S, E> by decider {}
+suspend fun <C, S, E> StateStoredAggregate<C, S, E>.handle(command: C): S =
+    (command.fetchState() ?: initialState)
+        .computeNewState(command)
+        .save()
 
+/**
+ * Extension function - Handles the [Flow] of command messages of type [C]
+ *
+ * @param commands [Flow] of Command messages of type [C]
+ * @return [Flow] of State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <C, S, E> StateStoredAggregate<C, S, E>.handle(commands: Flow<C>): Flow<S> =
+    commands.map { handle(it) }
 
 /**
  * Extension function - Publishes the command of type [C] to the state stored aggregate of type  [StateStoredAggregate]<[C], [S], *>
@@ -213,6 +183,76 @@ suspend fun <C, S> C.publishTo(aggregate: StateStoredAggregate<C, S, *>): S = ag
  * @author Иван Дугалић / Ivan Dugalic / @idugalic
  */
 fun <C, S> Flow<C>.publishTo(aggregate: StateStoredAggregate<C, S, *>): Flow<S> = aggregate.handle(this)
+
+
+// #################################################
+// ##########   Arrow Kotlin Extensions   ##########
+// #################################################
+
+/**
+ * Extension function - Handles the command message of type [C]
+ *
+ * @param command Command message of type [C]
+ * @return Either [Error] or State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+suspend fun <C, S, E> StateStoredAggregate<C, S, E>.handleEither(command: C): Either<Error, S> {
+    /**
+     * Inner function - Computes new State based on the previous State and the [command] or fails.
+     *
+     * @param command of type [C]
+     * @return [Either] the newly computed state of type [S] or [Error]
+     */
+    suspend fun S.eitherComputeNewStateOrFail(command: C): Either<Error, S> =
+        Either.catch {
+            computeNewState(command)
+        }.mapLeft { throwable ->
+            Error.CalculatingNewStateFailed(this, command, throwable)
+        }
+
+    /**
+     * Inner function - Fetch state - either version
+     *
+     * @receiver Command of type [C]
+     * @return [Either] [Error] or the State of type [S]?
+     */
+    suspend fun C.eitherFetchStateOrFail(): Either<Error.FetchingStateFailed<C>, S?> =
+        Either.catch {
+            fetchState()
+        }.mapLeft { throwable -> Error.FetchingStateFailed(this, throwable) }
+
+    /**
+     * Inner function - Save state - either version
+     *
+     * @receiver State of type [S]
+     * @return [Either] [Error] or the newly saved State of type [S]
+     */
+    suspend fun S.eitherSaveOrFail(): Either<Error.StoringStateFailed<S>, S> =
+        Either.catch {
+            save()
+        }.mapLeft { throwable -> Error.StoringStateFailed(this, throwable) }
+
+    // Arrow provides a Monad instance for Either. Except for the types signatures, our program remains unchanged when we compute over Either. All values on the left side assume to be Right biased and, whenever a Left value is found, the computation short-circuits, producing a result that is compatible with the function type signature.
+    return either {
+        (command.eitherFetchStateOrFail().bind() ?: initialState)
+            .eitherComputeNewStateOrFail(command).bind()
+            .eitherSaveOrFail().bind()
+    }
+}
+
+/**
+ * Extension function - Handles the [Flow] of command messages of type [C]
+ *
+ * @param commands [Flow] of Command messages of type [C]
+ * @return [Flow] of [Either] [Error] or State of type [S]
+ *
+ * @author Иван Дугалић / Ivan Dugalic / @idugalic
+ */
+fun <C, S, E> StateStoredAggregate<C, S, E>.handleEither(commands: Flow<C>): Flow<Either<Error, S>> =
+    commands
+        .map { handleEither(it) }
+        .catch { emit(Either.Left(Error.CommandPublishingFailed(it))) }
 
 /**
  * Extension function - Publishes the command of type [C] to the state stored aggregate of type  [StateStoredAggregate]<[C], [S], *>
